backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic[email]
python-multipart
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-dotenv
PyPDF2
apscheduler
sqlalchemy
pymysql